Refactor lots parsing and add Seller type
Moved lots parsing logic from pkg/parse/lots.go to pkg/scraper/webpage/lots.go and enhanced it to extract detailed seller information. Introduced a new Seller type in pkg/types/seller.go and updated Lot struct to use Seller instead of userLink/userName. Adjusted price parsing to use float32. Removed obsolete selectors.go and parse/lots.go files.

+++ pkg/types/price.go
@@ -1,6 +1,6 @@
 package types
 
 type Price struct {
-	Amount   float64 `json:"amount"`
+	Amount   float32 `json:"amount"`
 	Currency string  `json:"currency"`
 }
